update terakhir 11 juni

+++ app/Http/Controllers/PelangganController.php
@@ -2,9 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Http\Request;
 use App\Models\Kartu;
 use App\Models\Pelanggan;
-use Illuminate\Http\Request;
 
 class PelangganController extends Controller
 {
@@ -13,9 +13,9 @@ class PelangganController extends Controller
      */
     public function index()
     {
+        //
         $pelanggan = Pelanggan::all();
-
-        return view('admin.pelanggan.index', compact('pelanggan'));
+        return view ('admin.pelanggan.index', compact('pelanggan'));
     }
 
     /**
@@ -23,10 +23,10 @@ class PelangganController extends Controller
      */
     public function create()
     {
+        //
         $kartu = Kartu::all();
         $gender = ['L', 'P'];
-
-        return view('admin.pelanggan.create', compact('kartu', 'gender'));
+        return view ('admin.pelanggan.create', compact('kartu', 'gender'));
     }
 
     /**
@@ -34,28 +34,27 @@ class PelangganController extends Controller
      */
     public function store(Request $request)
     {
-        // proses upload foto
-        // jika file foto ada yang terupload
-        if (!empty($request->foto)) {
-            // maka proses berikut yang dijalankan
+         //proses upload foto
+        //jika file foto ada yang terupload
+        if(!empty($request->foto)){
+            //maka proses berikut yang dijalankan
             $fileName = 'foto-'.uniqid().'.'.$request->foto->extension();
-            // setelah tau fotonya sudah masuk maka tempatkan ke public
+            //setelah tau fotonya sudah masuk maka tempatkan ke public
             $request->foto->move(public_path('admin/image'), $fileName);
         } else {
             $fileName = '';
         }
-        // tambah data menggunakan eloquent
-        $pelanggan = new Pelanggan();
+        //tambah data menggunakan eloquent
+        $pelanggan = new Pelanggan;
         $pelanggan->kode = $request->kode;
         $pelanggan->nama = $request->nama;
         $pelanggan->jk = $request->jk;
         $pelanggan->tmp_lahir = $request->tmp_lahir;
         $pelanggan->tgl_lahir = $request->tgl_lahir;
-        $pelanggan->email = $request->email;
+        $pelanggan->email =$request->email;
         $pelanggan->foto = $fileName;
         $pelanggan->kartu_id = $request->kartu_id;
         $pelanggan->save();
-
         return redirect('admin/pelanggan');
     }
 
@@ -64,8 +63,9 @@ class PelangganController extends Controller
      */
     public function show(string $id)
     {
+        //show eloquent
         $pelanggan = Pelanggan::find($id);
-
+        // dd($pelanggan);
         return view('admin.pelanggan.detail', compact('pelanggan'));
     }
 
@@ -74,11 +74,11 @@ class PelangganController extends Controller
      */
     public function edit(string $id)
     {
+        //
         $pl = Pelanggan::find($id);
         $kartu = Kartu::all();
         $gender = ['L', 'P'];
-
-        return view('admin.pelanggan.edit', compact('pl', 'kartu', 'gender'));
+        return view ('admin.pelanggan.edit', compact('pl','kartu', 'gender'));
     }
 
     /**
@@ -86,36 +86,34 @@ class PelangganController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        // foto lama
-        $fotoLama = Pelanggan::select('foto')->where('id', $id)->get();
-        foreach ($fotoLama as $f1) {
-            $fotoLama = $f1->foto;
-        }
-        // jika foto sudah ada yang terupload
-        if (!empty($request->foto)) {
-            // maka proses selanjutnya
-            if (!empty($fotoLama->foto)) {
-                unlink(public_path('admin/image'.$fotoLama->foto));
-            }
-            // proses ganti foto
-            $fileName = 'foto-'.$request->id.'.'.$request->foto->extension();
-            // setelah tau fotonya sudah masuk maka tempatkan ke public
-            $request->foto->move(public_path('admin/image'), $fileName);
-        } else {
-            $fileName = $fotoLama;
-        }
-        // tambah data menggunakan eloquent
+        //
+       //foto lama 
+       $fotoLama = Pelanggan::select('foto')->where('id', $id)->get();
+       foreach($fotoLama as $f1){
+           $fotoLama = $f1->foto;
+       }
+       //jika foto sudah ada yang terupload 
+       if(!empty($request->foto)){
+           //maka proses selanjutnya 
+       if(!empty($fotoLama->foto)) unlink(public_path('admin/image'.$fotoLama->foto));
+       //proses ganti foto
+           $fileName = 'foto-'.$request->id.'.'.$request->foto->extension();
+           //setelah tau fotonya sudah masuk maka tempatkan ke public
+           $request->foto->move(public_path('admin/image'), $fileName);
+       } else{
+           $fileName = $fotoLama;
+       }
+        //tambah data menggunakan eloquent
         $pelanggan = Pelanggan::find($id);
         $pelanggan->kode = $request->kode;
         $pelanggan->nama = $request->nama;
         $pelanggan->jk = $request->jk;
         $pelanggan->tmp_lahir = $request->tmp_lahir;
         $pelanggan->tgl_lahir = $request->tgl_lahir;
-        $pelanggan->email = $request->email;
+        $pelanggan->email =$request->email;
         $pelanggan->foto = $fileName;
         $pelanggan->kartu_id = $request->kartu_id;
         $pelanggan->save();
-
         return redirect('admin/pelanggan');
     }
 
@@ -124,6 +122,7 @@ class PelangganController extends Controller
      */
     public function destroy(string $id)
     {
+        //
         $pelanggan = Pelanggan::find($id);
         $pelanggan->delete();
